(#18) Move and make internal classes that don't need to be seen from outside
