+++ src/Stravaig.Configuration.SqlServer.Tests/SourceBuilderTests.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using Microsoft.Extensions.Configuration;
 using NUnit.Framework;
 using Shouldly;
+using Stravaig.Configuration.SqlServer.Glue;
 
 namespace Stravaig.Configuration.SqlServer.Tests;
