+++ src/Stravaig.Configuration.SqlServer/SqlServerConfigurationBuilderExtensions.cs
@@ -1,5 +1,6 @@
 using System;
 using Microsoft.Extensions.Configuration;
+using Stravaig.Configuration.SqlServer.Glue;
 
 namespace Stravaig.Configuration.SqlServer;
