+++ src/Stravaig.Configuration.SqlServer/SqlServerConfigurationOptions.cs
@@ -1,4 +1,5 @@
 using System;
+using Stravaig.Configuration.SqlServer.Glue;
 
 namespace Stravaig.Configuration.SqlServer;
